+++ src/emmental/logging/logging_manager.py
@@ -6,6 +6,7 @@ from torch.optim.lr_scheduler import _LRScheduler
 from torch.optim.optimizer import Optimizer
 
 from emmental.logging.checkpointer import Checkpointer
+from emmental.logging.json_writer import JsonWriter
 from emmental.logging.log_writer import LogWriter
 from emmental.logging.tensorboard_writer import TensorBoardWriter
 from emmental.logging.wandb_writer import WandbWriter
@@ -107,9 +108,9 @@ class LoggingManager(object):
         writer_opt = Meta.config["logging_config"]["writer_config"]["writer"]
 
         if writer_opt is None:
-            self.writer = None
-        elif writer_opt == "json":
             self.writer = LogWriter()
+        elif writer_opt == "json":
+            self.writer = JsonWriter()
         elif writer_opt == "tensorboard":
             self.writer = TensorBoardWriter()
         elif writer_opt == "wandb":