+++ src/emmental/logging/tensorboard_writer.py
@@ -1,11 +1,13 @@
 """Emmental tensor board writer."""
+import copy
 import json
-from typing import Dict, Union
+from typing import Union
 
 from torch.utils.tensorboard import SummaryWriter
 
 from emmental.logging.log_writer import LogWriter
 from emmental.meta import Meta
+from emmental.utils.utils import convert_to_serializable_json
 
 
 class TensorBoardWriter(LogWriter):
@@ -15,21 +17,9 @@ class TensorBoardWriter(LogWriter):
         """Initialize TensorBoardWriter."""
         super().__init__()
 
-        # Set up tensorboard summary writer
+        # Set up tensorboard summary writer and save config
         self.writer = SummaryWriter(Meta.log_path)
         self.write_config()
-
-    def add_scalar_dict(
-        self, metric_dict: Dict[str, Union[float, int]], step: Union[float, int]
-    ) -> None:
-        """Log a scalar variable.
-
-        Args:
-          metric_dict: The metric dict.
-          step: The current step.
-        """
-        for name, value in metric_dict.items():
-            self.add_scalar(name, value, step)
 
     def add_scalar(
         self, name: str, value: Union[float, int], step: Union[float, int]
@@ -49,11 +39,8 @@ class TensorBoardWriter(LogWriter):
         Args:
           config_filename: The config filename, defaults to "config.yaml".
         """
-        try:
-            config = json.dumps(Meta.config)
-            self.writer.add_text(tag="config", text_string=config)
-        except TypeError:
-            pass
+        config = json.dumps(convert_to_serializable_json(copy.deepcopy(Meta.config)))
+        self.writer.add_text(tag="config", text_string=config)
         super().write_config(config_filename)
 
     def write_log(self, log_filename: str = "log.json") -> None: