+++ tests/logging/test_logging_manager.py
@@ -4,6 +4,7 @@ import logging
 import pytest
 
 import emmental
+from emmental.logging.json_writer import JsonWriter
 from emmental.logging.log_writer import LogWriter
 from emmental.logging.logging_manager import LoggingManager
 from emmental.logging.tensorboard_writer import TensorBoardWriter
@@ -205,7 +206,7 @@ def test_logging_manager_json(caplog):
 
     logging_manager.update(5)
 
-    assert type(logging_manager.writer) == LogWriter
+    assert type(logging_manager.writer) == JsonWriter
 
 
 def test_logging_manager_tensorboard(caplog):
@@ -253,7 +254,7 @@ def test_logging_manager_no_writer(caplog):
 
     logging_manager.update(5)
 
-    assert logging_manager.writer is None
+    assert type(logging_manager.writer) == LogWriter
 
 
 def test_logging_manager_wrong_writer(caplog):